Updated playerLeveling

+++ src/main/java/nl/grapjeje/opengrinding/utils/configuration/LevelConfig.java
@@ -3,5 +3,5 @@ package nl.grapjeje.opengrinding.utils.configuration;
 public interface LevelConfig {
     int getMaxLevel();
     double getXpForLevel(int level);
-    Integer getLevelOverride(int level);
+    Double getLevelOverride(int level);
 }
